Rails-Site --  Essentials, SignUp
The following commit is the baseline to build off of. It includes the essentials (in regard to Gems, run-dependent files) as well as the modeling necessary to sign users up to the site. To protect and secure user information being sent over the network, SecureSSL was enabled in *config/environments/production.rb*. Puma was installed as the sites WebServer to handle multiple incomming requests since our userbase will be uploading at numerous times throughout the day.